WIP #594
