+++ apstools/_utils/__init__.py
@@ -1,14 +1,16 @@
-__all__ = """
-    getDefaultNamespace
-    ipython_profile_name
-    ipython_shell_namespace
-    listdevice
-    listplans
-    OverrideParameters
-""".split()
-
+from .catalog import db_query
+from .catalog import findCatalogsInNamespace
+from .catalog import getCatalog
+from .catalog import getDatabase
+from .catalog import getDefaultCatalog
+from .catalog import getDefaultDatabase
 from .device_info import listdevice
 from .list_plans import listplans
+from .list_runs import getRunData
+from .list_runs import getRunDataValue
+from .list_runs import listruns
+from .list_runs import ListRuns
+from .list_runs import summarize_runs
 from .override_parameters import OverrideParameters
 from .profile_support import getDefaultNamespace
 from .profile_support import ipython_profile_name
