+++ apstools/_utils/override_parameters.py
@@ -138,7 +138,6 @@ class OverrideParameters:
             values.append(self.pick(parm, "--undefined--"))
         return pd.DataFrame(dict(parameter=parameters, value=values))
 
-
 # -----------------------------------------------------------------------------
 # :author:    Pete R. Jemian
 # :email:     jemian@anl.gov
